Conflicts:
	gnuradio-core/src/lib/general/general.i
	gnuradio-core/src/lib/general/gr_pll_carriertracking_cc.cc
	gnuradio-core/src/lib/general/gr_pll_freqdet_cf.cc
	gnuradio-core/src/lib/general/gr_pll_refout_cc.cc
	gnuradio-core/src/python/gnuradio/gr/qa_pll_carriertracking.py
	gnuradio-core/src/python/gnuradio/gr/qa_pll_freqdet.py
	gnuradio-core/src/python/gnuradio/gr/qa_pll_refout.py
	gr-digital/lib/digital_constellation_receiver_cb.cc
	gr-digital/python/Makefile.am
	gr-digital/python/__init__.py
	gr-digital/python/generic_mod_demod.py
	gr-digital/python/pkt.py
	gr-digital/python/psk2.py
	gr-digital/python/qam.py

+++ gr-digital/include/digital_constellation.h
@@ -85,7 +85,7 @@ public:
   unsigned int dimensionality() {return d_dimensionality;}
 
   unsigned int bits_per_symbol () {
-    return floor(log(d_constellation.size())/d_dimensionality/log(2));
+    return floor(log(double(d_constellation.size()))/d_dimensionality/log(2.0));
   }
   
   unsigned int arity () {
